+++ readthedocs/oauth/tasks.py
@@ -6,7 +6,7 @@ from djcelery import celery as celery_app
 from readthedocs.core.utils.tasks import PublicTask
 from readthedocs.core.utils.tasks import permission_check
 from readthedocs.core.utils.tasks import user_id_matches
-from .utils import services
+from .services import registry
 
 
 @permission_check(user_id_matches)
@@ -16,7 +16,7 @@ class SyncRemoteRepositories(PublicTask):
 
     def run_public(self, user_id):
         user = User.objects.get(pk=user_id)
-        for service_cls in services:
+        for service_cls in registry:
             service = service_cls.for_user(user)
             if service is not None:
                 service.sync(sync=True)
